Add a project-router file
In this commit, I added a project-router.js file to the project. This file is used
to handle GET / and POST / requests to our projects api resource. With the GET / endpoint, weare able to fetch the list of all projects while the POST / endpoint helps us add a new projects data respectively